// front-end/src/app/models/generated/models/ad.ts
/**
 * Ads API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Comment } from './comment';
import { Photo } from './photo';
import { Contact } from './contact';
import { Location } from './location';


export interface Ad { 
    id?: string | null;
    title?: string | null;
    body?: string | null;
    category?: string | null;
    price?: number | null;
    tags?: Array<string> | null;
    location?: Location;
    contact?: Contact;
    comments?: Array<Comment> | null;
    photos?: Array<Photo> | null;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string;
}